item-tree restructuring
